refactor: separate file for gallery panel
